+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 % erl-cmdline changelog
 
 # Next Version
+## Features
+- Add `cmdline:program_name/1`.
 ## Misc
 - Introduce `cmdline:config/0` as alias for `cmdline_config:config/0`.